pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-fleetpings"
description = "Fleet Ping Tool for Alliance Auth supporting pings via webhooks to Discord."
readme = "README.md"
keywords = [
    "allianceauth",
    "eveonline",
    "fleetpings",
]
license = {file = "LICENSE"}
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = ">=3.8"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.0.0b1",
    "allianceauth-app-utils>=1.13",
    "dhooks-lite>=0.6.1",
]
[project.optional-dependencies]
tests-allianceauth-latest = [
    "aa-srp>=1.21",
    "aa-timezones",
    "coverage",
    "django-eveuniverse",
    "django-webtest",
    "Faker",
    "fittings",
]
[project.urls]
Changelog = "https://github.com/ppfeufer/aa-fleetpings/blob/master/CHANGELOG.md"
Documentation = "https://github.com/ppfeufer/aa-fleetpings/blob/master/README.md"
Donations = "https://ko-fi.com/ppfeufer"
Homepage = "https://github.com/ppfeufer/aa-fleetpings"
Source = "https://github.com/ppfeufer/aa-fleetpings.git"
Tracker = "https://github.com/ppfeufer/aa-fleetpings/issues"

[tool.hatch.build]
include = [
    "/fleetpings",
]

[tool.hatch.version]
path = "fleetpings/__init__.py"
